# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 'NSI-IT-LMS API Documentation'
  description: 'Complete, production-ready REST API documentation for the NSI IT Learning Management System (Nityashree Infosystems LMS). Includes Authentication, Role-based User Management, Courses, Modules, Lectures & Notes, Cohort Batches & Allocations, Quizzes & Judge0 Code Execution, Schedules, Reviews, Announcements, Notifications, Platform Settings, and Reporting Hub.'
  version: 1.0.0
  contact:
    name: 'NSI IT Support'
    email: support@nsiit.com
servers:
  -
    url: 'http://localhost:5000'
    description: 'Local Development Server'
...
paths:
...
  /api/quizzes/languages:
    get:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Get supported Judge0 programming languages'
      security:
        -
          bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  languages: {type: array, items: {type: string}}
  /api/quizzes/sample-csv:
    get:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Download sample CSV template for quiz authoring'
      security:
        -
          bearerAuth: []
      responses:
        '200':
          content:
            text/csv:
              schema:
                type: string
                example: 'question_text,question_type,marks,difficulty,option_a,option_b,correct_option...'
  /api/quizzes/import-csv:
    post:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Create new test / assessment by uploading or parsing CSV'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - course_id
                - module_id
              properties:
                title:
                  type: string
                  example: 'Python Fundamentals & OOP Assessment'
                description:
                  type: string
                  example: 'Assessment imported via CSV'
                course_id:
                  type: integer
                  example: 1
                module_id:
                  type: integer
                  example: 1
                duration_minutes:
                  type: integer
                  example: 30
                passing_marks:
                  type: number
                  example: 15
                max_attempts:
                  type: integer
                  example: 2
                csv_content:
                  type: string
                  description: 'Raw CSV text with question rows'
                questions:
                  type: array
                  items: {type: object}
                  description: 'Pre-parsed question objects'
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  message: {type: string}
                  data: {$ref: '#/components/schemas/Quiz'}
  '/api/quizzes/{quizId}/import-csv':
    post:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Bulk import questions from CSV into an existing quiz'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                csv_content:
                  type: string
                  description: 'Raw CSV text'
                questions:
                  type: array
                  items: {type: object}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  message: {type: string}
                  data: {$ref: '#/components/schemas/Quiz'}
  /api/quizzes:
    get:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'List all quizzes'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: course_id
          in: query
          schema:
            type: integer
        -
          name: module_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  quizzes: {type: array, items: {$ref: '#/components/schemas/Quiz'}}
    post:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Create new quiz'
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: 'JavaScript Promises Quiz'
                description:
                  type: string
                  example: 'Async/await and microtasks'
                course_id:
                  type: integer
                  example: 1
                module_id:
                  type: integer
                  example: 1
                duration_minutes:
                  type: integer
                  example: 30
                total_marks:
                  type: number
                  example: 20
                passing_marks:
                  type: number
                  example: 10
                max_attempts:
                  type: integer
                  example: 2
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/quizzes/{quizId}':
    get:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Get quiz with full questions and test cases'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  quiz: {$ref: '#/components/schemas/Quiz'}
    put:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Update quiz settings'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                duration_minutes:
                  type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Delete quiz'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/quizzes/{quizId}/publish':
    patch:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Publish quiz to students'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/quizzes/{quizId}/close':
    patch:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Close quiz for submissions'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/quizzes/{quizId}/questions':
    post:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Add question to quiz (MCQ or CODING)'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question_type
                - question_text
              properties:
                question_type:
                  type: string
                  enum: [MCQ, CODING]
                  example: MCQ
                question_text:
                  type: string
                  example: 'Which method schedules a macro-task?'
                marks:
                  type: number
                  example: 5
                difficulty:
                  type: string
                  enum: [EASY, MEDIUM, HARD]
                  example: EASY
                programming_language:
                  type: string
                  example: javascript
                starter_code:
                  type: string
                  example: "function solution() {\n\n}"
                expected_output:
                  type: string
                  example: 'Hello World'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/questions/{questionId}':
    put:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Update question details'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question_text:
                  type: string
                marks:
                  type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Delete question'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/questions/{questionId}/options':
    post:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Add option choice to MCQ question'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - option_label
                - option_text
                - is_correct
              properties:
                option_label:
                  type: string
                  example: A
                option_text:
                  type: string
                  example: setTimeout
                is_correct:
                  type: boolean
                  example: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/options/{optionId}':
    put:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Update option choice'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: optionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                option_text:
                  type: string
                is_correct:
                  type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Delete option choice'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: optionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  '/api/quizzes/{quizId}/attempts':
    get:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Get student submissions and attempts for a quiz'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  attempts: {type: array, items: {$ref: '#/components/schemas/QuizAttempt'}}
  '/api/quizzes/attempts/{attemptId}':
    get:
      tags:
        - 'Quizzes (Admin & Instructor)'
      summary: 'Get detailed attempt report with student answers and score breakdown'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: attemptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean}
                  attempt: {$ref: '#/components/schemas/QuizAttempt'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter standard JWT Bearer token: Bearer <token>'
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Operation completed successfully'
...
    Quiz:
      type: object
      properties:
        id:
          type: integer
          example: 1
        session_id:
          type: integer
          nullable: true
          example: null
        module_id:
          type: integer
          nullable: true
          example: 1
        course_id:
          type: integer
          nullable: true
          example: 1
        title:
          type: string
          example: 'React Hooks & State Test'
        description:
          type: string
          example: 'Assess understanding of useState and useEffect'
        duration_minutes:
          type: integer
          example: 45
        total_marks:
          type: number
          example: 50
        passing_marks:
          type: number
          example: 20
        max_attempts:
          type: integer
          example: 2
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - CLOSED
            - CANCELLED
          example: PUBLISHED
...
    QuizAttempt:
      type: object
      properties:
        id:
          type: integer
          example: 1
        quiz_id:
          type: integer
          example: 1
        student_id:
          type: integer
          example: 3
        attempt_number:
          type: integer
          example: 1
        status:
          type: string
          enum:
            - IN_PROGRESS
            - SUBMITTED
            - AUTO_SUBMITTED
            - CANCELLED
          example: SUBMITTED
        score:
          type: number
          example: 45
        total_marks:
          type: number
          example: 50
        passed:
          type: boolean
          example: true
        started_at:
          type: string
          format: date-time
        submitted_at:
          type: string
          format: date-time
          nullable: true
...